requirements.txt
PySide6>=6.5
requests>=2.31
beautifulsoup4>=4.12
Pillow>=10.0
